test(drush): add cache rebuild test
