+++ composer.json
@@ -16,12 +16,12 @@
             "Ekino\\Drupal\\Debug\\Tests\\Integration\\": "tests/Integration/",
             "Ekino\\Drupal\\Debug\\Tests\\Traits\\": "tests/Traits/",
             "Ekino\\Drupal\\Debug\\Tests\\Unit\\": "tests/Unit/src/",
-            "Drupal\\hit_dynamic_page_cache\\": "tests/Integration/DisableDynamicPageCache/fixtures/modules/hit_dynamic_page_cache/src/",
-            "Drupal\\hit_page_cache\\": "tests/Integration/DisableInternalPageCache/fixtures/modules/hit_page_cache/src/",
-            "Drupal\\hit_render_cache\\": "tests/Integration/DisableRenderCache/fixtures/modules/hit_render_cache/src/",
-            "Drupal\\use_custom_route\\": "tests/Integration/WatchRoutingDefinitions/fixtures/modules/use_custom_route/src/",
-            "Drupal\\use_custom_service\\": "tests/Integration/WatchContainerDefinitions/fixtures/modules/use_custom_service/src/",
-            "Drupal\\use_twig_template\\": "tests/Integration/DisableTwigCache/fixtures/modules/use_twig_template/src/"
+            "Drupal\\hit_dynamic_page_cache\\": "tests/Integration/Action/DisableDynamicPageCache/fixtures/modules/hit_dynamic_page_cache/src/",
+            "Drupal\\hit_page_cache\\": "tests/Integration/Action/DisableInternalPageCache/fixtures/modules/hit_page_cache/src/",
+            "Drupal\\hit_render_cache\\": "tests/Integration/Action/DisableRenderCache/fixtures/modules/hit_render_cache/src/",
+            "Drupal\\use_custom_route\\": "tests/Integration/Action/WatchRoutingDefinitions/fixtures/modules/use_custom_route/src/",
+            "Drupal\\use_custom_service\\": "tests/Integration/Action/WatchContainerDefinitions/fixtures/modules/use_custom_service/src/",
+            "Drupal\\use_twig_template\\": "tests/Integration/Action/DisableTwigCache/fixtures/modules/use_twig_template/src/"
         },
         "classmap": [
             "tests/Unit/src/Kernel/test_classes/"
@@ -49,6 +49,7 @@
     "require-dev": {
         "composer/composer": "^1.7",
         "drupal-composer/drupal-scaffold": "^2.5",
+        "drush/drush": "^9.0",
         "fabpot/goutte": "^3.2.3",
         "friendsofphp/php-cs-fixer": "^2.14",
         "guzzlehttp/guzzle": "^6.3",
