+++ phpunit.xml.dist
@@ -16,13 +16,16 @@
         <testsuite name="unit">
             <directory suffix="Test.php">tests/Unit</directory>
         </testsuite>
-        <testsuite name="integration">
-            <directory suffix="Test.php">tests/Integration</directory>
+        <testsuite name="integration/action">
+            <directory suffix="Test.php">tests/Integration/Action</directory>
+        </testsuite>
+        <testsuite name="integration/drush">
+            <directory suffix="Test.php">tests/Integration/Drush</directory>
         </testsuite>
     </testsuites>
 
     <listeners>
-        <listener class="Ekino\Drupal\Debug\Tests\Integration\SetupListener" file="tests/Integration/SetupListener.php" />
+        <listener class="Ekino\Drupal\Debug\Tests\Integration\Action\SetupListener" file="tests/Integration/Action/SetupListener.php" />
     </listeners>
 
     <filter>
