+++ tests/Integration/Action/DisableCSSAggregation/DisableCSSAggregationActionTest.php
@@ -11,19 +11,19 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Ekino\Drupal\Debug\Tests\Integration\DisableJSAggregation;
+namespace Ekino\Drupal\Debug\Tests\Integration\Action\DisableCSSAggregation;
 
-use Ekino\Drupal\Debug\Tests\Integration\AbstractTestCase;
+use Ekino\Drupal\Debug\Tests\Integration\Action\AbstractActionTestCase;
 use Symfony\Component\BrowserKit\Client;
 
-class DisableJSAggregationActionTest extends AbstractTestCase
+class DisableCSSAggregationActionTest extends AbstractActionTestCase
 {
     /**
      * {@inheritdoc}
      */
     protected function doTestInitialBehaviorWithDrupalKernel(Client $client): void
     {
-        $this->assertSame(1, $this->countScripts($client));
+        $this->assertSame(1, $this->countStylesheetLinks($client));
     }
 
     /**
@@ -31,14 +31,14 @@ class DisableJSAggregationActionTest extends AbstractTestCase
      */
     protected function doTestTargetedBehaviorWithDebugKernel(Client $client): void
     {
-        $this->assertSame(2, $this->countScripts($client));
+        $this->assertGreaterThan(10, $this->countStylesheetLinks($client));
     }
 
     /**
      * @return int
      */
-    private function countScripts(Client $client): int
+    private function countStylesheetLinks(Client $client): int
     {
-        return \iterator_count($client->request('GET', '/')->filterXPath('descendant-or-self::script'));
+        return \iterator_count($client->request('GET', '/')->filterXPath('descendant-or-self::link[@rel="stylesheet"]'));
     }
 }
