+++ tests/Integration/Action/DisableDynamicPageCache/fixtures/ControllerTemplate.php
