+++ tests/Integration/Action/DisableDynamicPageCache/fixtures/modules/hit_dynamic_page_cache/hit_dynamic_page_cache.info.yml
