+++ tests/Integration/Action/DisableInternalPageCache/fixtures/ControllerTemplate.php
