+++ tests/Integration/Action/DisableInternalPageCache/fixtures/modules/hit_page_cache/hit_page_cache.info.yml
