+++ tests/Integration/Action/DisableJSAggregation/fixtures/modules/add_js_scripts/add_js_scripts.libraries.yml
