+++ tests/Integration/Action/DisableRenderCache/fixtures/ControllerTemplate.php
