+++ tests/Integration/Action/DisableRenderCache/fixtures/modules/hit_render_cache/hit_render_cache.info.yml
