+++ tests/Integration/Action/DisplayDumpLocation/DisplayDumpLocationActionTest.php
@@ -11,12 +11,12 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Ekino\Drupal\Debug\Tests\Integration\DisplayDumpLocation;
+namespace Ekino\Drupal\Debug\Tests\Integration\Action\DisplayDumpLocation;
 
-use Ekino\Drupal\Debug\Tests\Integration\AbstractTestCase;
+use Ekino\Drupal\Debug\Tests\Integration\Action\AbstractActionTestCase;
 use Symfony\Component\BrowserKit\Client;
 
-class DisplayDumpLocationActionTest extends AbstractTestCase
+class DisplayDumpLocationActionTest extends AbstractActionTestCase
 {
     /**
      * {@inheritdoc}
