+++ tests/Integration/Action/DisplayDumpLocation/fixtures/modules/add_dump_die/add_dump_die.info.yml
