+++ tests/Integration/Action/DisplayPrettyExceptions/fixtures/modules/throw_uncaught_exception/throw_uncaught_exception.info.yml
