+++ tests/Integration/Action/DisplayPrettyExceptionsASAP/fixtures/modules/throw_uncaught_exception/throw_uncaught_exception.module
