+++ tests/Integration/Action/ThrowErrorsAsExceptions/fixtures/modules/trigger_error/trigger_error.info.yml
