+++ tests/Integration/Action/WatchContainerDefinitions/WatchContainerDefinitionsActionTest.php
@@ -11,14 +11,14 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Ekino\Drupal\Debug\Tests\Integration\WatchContainerDefinitions;
+namespace Ekino\Drupal\Debug\Tests\Integration\Action\WatchContainerDefinitions;
 
 use Carbon\Carbon;
-use Ekino\Drupal\Debug\Tests\Integration\AbstractTestCase;
+use Ekino\Drupal\Debug\Tests\Integration\Action\AbstractActionTestCase;
 use Ekino\Drupal\Debug\Tests\Traits\FileHelperTrait;
 use Symfony\Component\BrowserKit\Client;
 
-class WatchContainerDefinitionsActionTest extends AbstractTestCase
+class WatchContainerDefinitionsActionTest extends AbstractActionTestCase
 {
     use FileHelperTrait;
