+++ tests/Integration/Action/WatchContainerDefinitions/fixtures/ServiceProviderTemplate.php
