+++ tests/Integration/Action/WatchContainerDefinitions/fixtures/modules/use_custom_service/src/Service/FooService.php
