+++ tests/Integration/Action/WatchContainerDefinitions/fixtures/services_template.yml
