+++ tests/Integration/Action/WatchModulesHooksImplementations/fixtures/module_hook_to_add.php
