+++ tests/Integration/Action/WatchRoutingDefinitions/fixtures/modules/use_custom_route/src/Controller/FooController.php
