+++ tests/Integration/Action/WatchRoutingDefinitions/fixtures/modules/use_custom_route/use_custom_route.info.yml
