+++ tests/Integration/Action/WatchRoutingDefinitions/fixtures/routing_template.yml
